fix: migration

+++ packages/app/src/routes/api/v1/users/[userId]/+server.ts
@@ -1,7 +1,8 @@
 import type { NDKUserProfile } from '@nostr-dev-kit/ndk'
 import { error, json } from '@sveltejs/kit'
-import { authorize, authorizeUserless } from '$lib/auth'
+import { authorize, authorizeEditorOrAdmin, authorizeUserless } from '$lib/auth'
 import { usersFilterSchema } from '$lib/schema'
+import { decodeJwtToEvent } from '$lib/server/nostrAuth.service'
 import { createUser, deleteUser, getRichUsers, getUserById, updateUser, updateUserFromNostr, userExists } from '$lib/server/users.service'
 
 import type { RequestHandler } from './$types'
@@ -78,15 +79,34 @@ export const POST: RequestHandler = async ({ request, params: { userId } }) => {
 }
 
 export const DELETE: RequestHandler = async ({ params, request }) => {
+	const authorizationHeader = request.headers.get('Authorization')
 	const { userId } = params
 
-	try {
-		await authorize(request, userId, 'DELETE')
-		return json(await deleteUser(userId))
-	} catch (e) {
-		if (e.status) {
-			error(e.status, e.message)
+	if (!authorizationHeader) {
+		throw error(401, 'Authorization header missing')
+	}
+
+	const token = decodeJwtToEvent(authorizationHeader)
+
+	console.log(token.pubkey, userId)
+
+	if (token.pubkey !== userId) {
+		try {
+			await authorizeEditorOrAdmin(request, 'DELETE')
+			console.log('authorized')
+			return json(await deleteUser(userId))
+		} catch (e) {
+			throw error(401, 'Invalid Token')
 		}
-		error(500, JSON.stringify(e))
+	} else {
+		try {
+			await authorize(request, userId, 'DELETE')
+			return json(await deleteUser(userId))
+		} catch (e) {
+			if (e.status) {
+				error(e.status, e.message)
+			}
+			throw error(500, JSON.stringify(e))
+		}
 	}
 }
